requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.1
django-formtools==2.5.1
django-js-asset==2.2.0
django-quill-editor==0.1.40
django-tinymce==4.0.0
idna==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tzdata==2024.1
urllib3==2.2.1
